test: assert the silver mapping, and fix the overflow it found
Nineteen tests over the adversarial fixture: the projection matches the DDL
column for column, the valid/quarantine split is a partition of the input,
first arrival wins and reverses when the offsets do, two id-less frames stay
two rows, and the MERGE text is pinned including the absent partition
predicate.

late_by_seconds becomes bigint. I had argued int was safe because
event_time_before_wikipedia bounds how late an event can be; it bounds lateness
in one direction only, and the column is computed before that rule runs. A
payload stamped 2099 gives -2,281,290,900 seconds, the narrowing cast raised
under ANSI mode, and the micro-batch would have died again on every replay.
Same poison pill as ADR-0018, found by the split test rather than in
production.

Every date assertion goes through date_format. A collected timestamp is
rendered in the JVM zone rather than the session zone, so on this machine a
naive .date() on 23:30Z reads as the following day.

+++ src/wikistream/streaming/silver.py
@@ -282,9 +282,13 @@ def to_candidates(frames: DataFrame) -> DataFrame:
         .withColumn("rev_old", F.col("_parsed.revision.old"))
         .withColumn("rev_new", F.col("_parsed.revision.new"))
         .withColumn("comment", F.col("_parsed.comment"))
+        # No cast. `unix_timestamp` returns a bigint and the column is a bigint, because
+        # a `.cast("int")` here raised under ANSI mode on a payload stamped 2099 —
+        # lateness is computed before the rule that rejects such a row, so the cast saw
+        # -2.28e9 and killed the batch. Matching the widths removes the failure mode.
         .withColumn(
             "late_by_seconds",
-            (F.unix_timestamp("ingested_at") - F.unix_timestamp("event_time")).cast("int"),
+            F.unix_timestamp("ingested_at") - F.unix_timestamp("event_time"),
         )
         .withColumn("event_date", F.to_date(F.col("event_time")))
         .withColumn("failed_at", F.col("ingested_at"))
